Delete unused vendored boxing layer core/boxing.py (Stage 2)
core/boxing.py is a vendoring artifact: numba-cuda-mlir (like numba-cuda) does
no PyObject boxing/unboxing on the MLIR device path. It is imported nowhere at
init; its only consumers are pythonapi.to_native_value/from_native_value (the
default box/unbox handlers), which are themselves only reached from the vendored
Lower object-mode path (not MLIRLower) - dead on the MLIR path. Every box/unbox
implementation in it built llvmlite IR.

Move the two fallback handlers (box_unsupported/unbox_unsupported) into
pythonapi.py so to_native_value/from_native_value still resolve a default, drop
the lazy boxing imports, and delete core/boxing.py. This also removes the
core/boxing -> listobj.ListInstance coupling, unblocking listobj cleanup.

Verified: GPU suite (--ignore=tests/benchmarks) unchanged at 2 failed
(pre-existing), 4197 passed, 334 xfailed.
